Add stats_mode=diff and diff_mode=rectangle
`stats_mode=diff` makes `palettegen` focus on the changes between frames instead of the whole frame.
Given that the whole frame (usually) consists mainly of a static white background, this is likely to improve image quality

`diff_mode=rectangle` makes the gif encoding crop to the bounding box of the changing parts. This again is likely to improve image quality because it avoids error diffusion to parts of the image that do not change. It can also greatly reduce file-size for GIFs where only a small section of the image changes between frames.

See this comment for more details: 

https://github.com/JuliaPlots/Plots.jl/issues/423#issuecomment-342173040