Use `even` to tweak length of `make_steps` output
If `even` in false, `make_steps` returns an odd-length (2n-1) Vector as
it did before.  If even is `true`, `make_steps` now returns an
even-length (2n) Vector whose first (2n-1) elements are the same as
before and whose final element is the same as the final element of the
input Vector.