GR: Fix zscale support in surface plots (#5063)
Previously, `zscale` was being set to a Bool value (the result of the
set membership check). This caused a crash when querying the
`gr_z_log_scales` tuple on the next line, because that array is indexed
by symbols like :log10.

Given that the lines above assign the actual symbols to `xscale` and
`yscale`, it seems likely that `zscale` should behave the same.

Fixes: https://github.com/JuliaPlots/Plots.jl/issues/5062